backend/requirements.txt
django
django-cors-headers
django-summernote
djangorestframework
Pillow
psycopg2-binary
